% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_diar_ors.R
\name{ch_diar_ors}
\alias{ch_diar_ors}
\title{CH_DIAT_C_ORS}
\usage{
ch_diar_ors(KRdata)
}
\arguments{
\item{KRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CH_DIAT_C_ORS
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
