% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_EVUM_W_MOD.R
\name{FP_EVUM_W_MOD}
\alias{FP_EVUM_W_MOD}
\title{AH_TOBC_W_OTH}
\usage{
FP_EVUM_W_MOD(Rdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
AH_TOBC_W_OTH
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
