% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hk_hiv_consl_anc.R
\name{hk_hiv_consl_anc}
\alias{hk_hiv_consl_anc}
\title{HA_CATH_W_NRS
#HK_TEST_CONSL_F  IRdata
Pregnant women tested for HIV during ANC not receiving results}
\usage{
hk_hiv_consl_anc(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_CATH_W_NRS
#HK_TEST_CONSL_F  IRdata
Pregnant women tested for HIV during ANC not receiving results
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_NADA_W_UNT",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_unmet_tot)
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
