% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nt_bf_start_1hr.R
\name{nt_bf_start_1hr}
\alias{nt_bf_start_1hr}
\title{CN_BRFI_C_1HR}
\usage{
nt_bf_start_1hr(KRdata)
}
\arguments{
\item{KRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CN_BRFI_C_1HR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
